Complete rewrite of render function to run on gpu instead. Thanks AI Zooming into a number shows what it is. It also follows the color scheme Zooming in now zooms into your cursor instead of the center of the screen Lots of changes on drawing. Color is now calculated on gpu instead of stored, which spares memory and is faster since the bottleneck for gpu is throughput instead of calculation Changed structure size to benefit of not needing padding

+++ src/modules/math_logic.cpp
@@ -13,7 +13,7 @@ void NumberGenerator::AddPoint(uint64_t p) {
   NumberPoint pt;
   pt.x = static_cast<float>(r * std::cos(theta));
   pt.y = static_cast<float>(-r * std::sin(theta));
-  pt.p = static_cast<float>(p);
+  pt.p = static_cast<uint32_t>(p);
 
   points.push_back(pt);
 }
